Sofia-AI/sofia_lite/requirements.txt
fastapi
uvicorn[standard]
python-multipart
langchain==0.2.*
langchain-openai
langchain-core
langchain-community
openai
google-cloud-firestore==2.12.0
google-cloud-vision==3.4.4
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2
google-auth-oauthlib
python-dotenv
twilio
elevenlabs
pytz==2025.2
python-dateutil==2.9.0.post0
pytest
websockets
httpx
psutil
# Enhanced Language Detection ML dependencies
langdetect
textblob
spacy
nltk 
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.16.4 